fix: some rexstan fixes

+++ lib/extension_points/article.php
@@ -7,7 +7,7 @@ class article
     use ep_trait;
 
     /**
-     * @var \rex_addon
+     * @var \rex_addon_interface
      */
     private static $addon;
 
@@ -22,43 +22,55 @@ class article
         /**
          * a new article has been added
          */
-        if(self::$addon->getConfig('article_added')) {
+        if(is_bool(self::$addon->getConfig('article_added')) && self::$addon->getConfig('article_added')) {
             $this->add('ART_ADDED', 'RexActivity\EP\article::message');
         }
 
         /**
          * a article has been updated
          */
-        if(self::$addon->getConfig('article_updated')) {
+        if(is_bool(self::$addon->getConfig('article_updated')) && self::$addon->getConfig('article_updated')) {
             $this->update('ART_UPDATED', 'RexActivity\EP\article::message');
         }
 
         /**
          * a article has been deleted
          */
-        if(self::$addon->getConfig('article_deleted')) {
+        if(is_bool(self::$addon->getConfig('article_deleted')) && self::$addon->getConfig('article_deleted')) {
             $this->delete('ART_DELETED', 'RexActivity\EP\article::message');
         }
 
         /**
          * the article status has been changed
          */
-        if(self::$addon->getConfig('article_status')) {
+        if(is_bool(self::$addon->getConfig('article_status')) && self::$addon->getConfig('article_status')) {
             $this->status('ART_STATUS', 'RexActivity\EP\article::message');
         }
     }
 
-    public static function message(array $params, string $type, $additionalParams = null): string {
-        $article = \rex_article::get($params['id']);
+    /**
+     * @param array<string> $params
+     * @param string $type
+     * @param array<string>|null $additionalParams
+     * @return string
+     */
+    public static function message(array $params, string $type, ?array $additionalParams = null): string {
+        $article = \rex_article::get((int) $params['id']);
 
         $message = '<strong>Article:</strong> ';
 
-        if($article) {
+        if($article !== null) {
+            $clang = $params['clang'];
+
+            if (isset($params['clang_id'])) {
+                $clang = $params['clang_id'];
+            }
+
             $message .= '<a href="' . \rex_url::backendController([
                     'page' => 'content/edit',
                     'article_id' => $article->getId(),
                     'category_id' =>  $article->getCategoryId(),
-                    'clang_id' => isset($params['clang_id']) ?: $params['clang'],
+                    'clang_id' => $clang,
                     'mode' => 'edit'
                 ]) . '">';
             $message .= $article->getName();
@@ -73,7 +85,10 @@ class article
 
         if(isset($additionalParams['type'])) {
             $message .= self::$addon->i18n('type_'.$additionalParams['type']);
-            $message .= self::getStatus($article->isOnline(), $additionalParams);
+
+            if($article !== null) {
+                $message .= self::getStatus($article->isOnline(), $additionalParams);
+            }
         }
         else {
             $message .= self::$addon->i18n('type_'.$type);
