+++ lib/extension_points/category.php
@@ -7,7 +7,7 @@ class category
     use ep_trait;
 
     /**
-     * @var \rex_addon
+     * @var \rex_addon_interface
      */
     private static $addon;
 
@@ -17,37 +17,43 @@ class category
         /**
          * a new category has been added
          */
-        if(self::$addon->getConfig('category_added')) {
+        if (is_bool(self::$addon->getConfig('category_added')) && self::$addon->getConfig('category_added')) {
             $this->add('CAT_ADDED', 'RexActivity\EP\category::message');
         }
 
         /**
          * a category has been updated
          */
-        if(self::$addon->getConfig('category_updated')) {
+        if (is_bool(self::$addon->getConfig('category_updated')) && self::$addon->getConfig('category_updated')) {
             $this->update('CAT_UPDATED', 'RexActivity\EP\category::message');
         }
 
         /**
          * a category has been deleted
          */
-        if(self::$addon->getConfig('category_deleted')) {
+        if (is_bool(self::$addon->getConfig('category_deleted')) && self::$addon->getConfig('category_deleted')) {
             $this->delete('CAT_DELETED', 'RexActivity\EP\category::message');
         }
 
         /**
          * the category status has been changed
          */
-        if(self::$addon->getConfig('category_status')) {
+        if (is_bool(self::$addon->getConfig('category_status')) && self::$addon->getConfig('category_status')) {
             $this->status('CAT_STATUS', 'RexActivity\EP\category::message');
         }
     }
 
-    public static function message(array $params, string $type, $additionalParams = null): string {
-        $category = \rex_category::get($params['id']);
+    /**
+     * @param array<string> $params
+     * @param string $type
+     * @param array<string>|null $additionalParams
+     * @return string
+     */
+    public static function message(array $params, string $type, ?array $additionalParams = null): string {
+        $category = \rex_category::get((int) $params['id']);
         $message = '<strong>Category:</strong> ';
 
-        if($type === 'delete' && !empty($params)) {
+        if($type === 'delete' && $params !== []) {
             $message .= $params['name'];
             $message .= ' [' . $params['id'] . ']';
         }
@@ -58,7 +64,11 @@ class category
                     'category_id' =>  $params['id'],
                     'clang_id' => $params['clang'],
                 ]) . '">';
-            $message .= $category->getName();
+
+            if($category !== null) {
+                $message .= $category->getName();
+            }
+
             $message .= '</a>';
         }
 
@@ -66,7 +76,10 @@ class category
 
         if(isset($additionalParams['type'])) {
             $message .= self::$addon->i18n('type_'.$additionalParams['type']);
-            $message .= self::getStatus($category->isOnline(), $additionalParams);
+
+            if($category !== null) {
+                $message .= self::getStatus($category->isOnline(), $additionalParams);
+            }
         }
         else {
             $message .= self::$addon->i18n('type_'.$type);
