+++ lib/extension_points/media.php
@@ -7,7 +7,7 @@ class media
     use ep_trait;
 
     /**
-     * @var \rex_addon
+     * @var \rex_addon_interface
      */
     private static $addon;
 
@@ -17,25 +17,30 @@ class media
         /**
          * new media has been added
          */
-        if(self::$addon->getConfig('media_added')) {
+        if (is_bool(self::$addon->getConfig('media_added')) && self::$addon->getConfig('media_added')) {
             $this->add('MEDIA_ADDED', 'RexActivity\EP\media::message');
         }
 
         /**
          * media has been updated
          */
-        if(self::$addon->getConfig('media_updated')) {
+        if (is_bool(self::$addon->getConfig('media_updated')) && self::$addon->getConfig('media_updated')) {
             $this->update('MEDIA_UPDATED', 'RexActivity\EP\media::message');
         }
 
         /**
          * media has been deleted
          */
-        if(self::$addon->getConfig('media_deleted')) {
+        if (is_bool(self::$addon->getConfig('media_deleted')) && self::$addon->getConfig('media_deleted')) {
             $this->delete('MEDIA_DELETED', 'RexActivity\EP\media::message');
         }
     }
 
+    /**
+     * @param array $params
+     * @param string $type
+     * @return string
+     */
     private function message(array $params, string $type): string {
         $message = '<strong>Media:</strong> ';
