+++ lib/extension_points/template.php
@@ -7,7 +7,7 @@ class template
     use ep_trait;
 
     /**
-     * @var \rex_addon
+     * @var \rex_addon_interface
      */
     private static $addon;
 
@@ -17,25 +17,30 @@ class template
         /**
          * a new template has been added
          */
-        if(self::$addon->getConfig('template_added')) {
+        if (is_bool(self::$addon->getConfig('template_added')) && self::$addon->getConfig('template_added')) {
             $this->add('TEMPLATE_ADDED', 'RexActivity\EP\template::message');
         }
 
         /**
          * a template has been updated
          */
-        if(self::$addon->getConfig('template_updated')) {
+        if (is_bool(self::$addon->getConfig('template_updated')) && self::$addon->getConfig('template_updated')) {
             $this->update('TEMPLATE_UPDATED', 'RexActivity\EP\template::message');
         }
 
         /**
          * a template has been deleted
          */
-        if(self::$addon->getConfig('template_deleted')) {
+        if (is_bool(self::$addon->getConfig('template_deleted')) && self::$addon->getConfig('template_deleted')) {
             $this->delete('TEMPLATE_DELETED', 'RexActivity\EP\template::message');
         }
     }
 
+    /**
+     * @param array $params
+     * @param string $type
+     * @return string
+     */
     public static function message(array $params, string $type): string {
         $message = '<strong>Template:</strong> ';
