+++ lib/extension_points/user.php
@@ -7,7 +7,7 @@ class user
     use ep_trait;
 
     /**
-     * @var \rex_addon
+     * @var \rex_addon_interface
      */
     private static $addon;
 
@@ -17,25 +17,30 @@ class user
         /**
          * new user has been added
          */
-        if(self::$addon->getConfig('user_added')) {
+        if (is_bool(self::$addon->getConfig('user_added')) && self::$addon->getConfig('user_added')) {
             $this->add('USER_ADDED', 'RexActivity\EP\user::message');
         }
 
         /**
          * user has been updated
          */
-        if(self::$addon->getConfig('user_updated')) {
-            $this->update('USER_UPDATED', 'RexActivity\EP\user::message');
+            if (is_bool(self::$addon->getConfig('user_updated')) && self::$addon->getConfig('user_updated')) {
+                $this->update('USER_UPDATED', 'RexActivity\EP\user::message');
         }
 
         /**
          * user has been deleted
          */
-        if(self::$addon->getConfig('user_deleted')) {
+        if (is_bool(self::$addon->getConfig('user_deleted')) && self::$addon->getConfig('user_deleted')) {
             $this->delete('USER_DELETED', 'RexActivity\EP\user::message');
         }
     }
 
+    /**
+     * @param array $params
+     * @param string $type
+     * @return string
+     */
     private function message(array $params, string $type): string {
         $message = '<strong>User:</strong> ';
         $message .= '<a href="' . \rex_url::backendController(['page' => 'users/users', 'user_id' => $params['user']->getId()]) . '" title="' . $params['user']->getName() . '">' . $params['user']->getName() . '</a>';
